requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
djangorestframework==3.10.2
gunicorn==19.9.0
joblib==0.13.2
nltk==3.4.4
numpy==1.17.0
psycopg2==2.8.3
pytz==2019.1
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
whitenoise==3.2
